backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.0.1
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pymongo==4.5.0
dnspython==2.4.2